Remove style wrapper
Instead of wrapping every style with a div and  thus, potentially messing up the
behaviour of the style the section identification css class is now added
directly to the class list of the style. This allows to use div styles in
conjunction with bootstrap classes such as col and row.