Replace user input node labels in PHP
When changing the node labels directly in the svg the node size and position is
not updated. This can lead to an ugly representation or worse to cropping of
text. In order to avoid this, the labels are now replaced in the code string (on
the server side) which is then passed to the mermaid parser.

Note that I introduced a model to seperate concerns.