- Add markdown data_config field
In this ***JSON*** field we can configure a data retrieve params from the DB, either `static` or `dynamic` data. Example:
 ```
 [
	{
		"type": "static|dynamic",
		"table": "table_name | #url_param1",
		"fields": [
			{
				"field_name": "name | #url_param2",
				"field_holder": "@field_1",
				"not_found_text": "my field was not found",
				"retrieve": "first | last | all"
			}
		]
	}
]
```
If the page supports parameters, then the parameter can be accessed with `#` and the name of the paramer. Example `#url_param_name`.

In order to inlcude the retrieved data in the markdown field, include the `field_holder` that wa defined in the markdown text.

We can access multiple tables by adding another element to the array. The retrieve data from the column can be: `first` entry, `last` entry or `all` entries (concatenated with ;)

Full example of json:

```
[{
		"type": "static",
		"table": "#code",
		"retrieve": "first",
		"fields": [{
				"field_name": "kon",
				"field_holder": "@kon",
				"not_found_text": "my field was not found"
			}, {
				"field_name": "kogn1",
				"field_holder": "@kogn",
				"not_found_text": "my field was not found"
			}
		]
	}, {
		"type": "static",
		"table": "sample_static",
		"retrieve": "last",
		"fields": [{
				"field_name": "trace4",
				"field_holder": "@trace4",
				"not_found_text": "my field was not found"
			}, {
				"field_name": "trace5",
				"field_holder": "@trace5",
				"not_found_text": "my field was not found"
			}
		]
	}, {
		"type": "dynamic",
		"table": "form",
		"retrieve": "all",
		"fields": [{
				"field_name": "first_name",
				"field_holder": "@first_name",
				"not_found_text": "my field was not found"
			}, {
				"field_name": "lasName",
				"field_holder": "@last_name",
				"not_found_text": "my field was not found"
			}
		]
	}
]
```

- add procedure get_form_data_for_user_with_filter